PY3 requires bytes-like object with os.write
Start the server using python3 and the `--pidfile` option with a valid
path, the server crashes because it fails to create the pidfile.

The problem is due to the `os.write` function of python3, it explicitly
requires a bytes-like object as text argument. As python2 is fine with
both string and bytes, it is better to use bytes so it is compatible
with the two versions.