Remove PID file when the server shutdowns
The PID file creation in `make_pid_file` requires the PID file does not
exists in order to create a new file. If it does exists, it prints an
error message and exits.

As the PID file is not removed when the server shutdowns, it is not
possible to restart it without manually deleting the existing PID file
on the file system.

As the system can arbitrary re-assign the very same PID of the server
once it has exited, the file should be deleted along with the process
termination.